Added Types Flag
Added types flag so that you may specify to rotate 
```
-t, --types strings           filter by credential type (comma separated) (default [certificate,ssh,rsa,password,user,value,json])
```

or as an example:
```
./carousel rotate --ignore-update-mode --expires-within=1y --types=certificate
```
and this would rotate certs that expire within 1 year. 

**Reason:**
Prior to this ability to specify certificates, just using
```
./carousel rotate --ignore-update-mode --expires-within=1y
```
would have rotated `passwords` that were not in need of rotation if they were created on a certain date.